FUNIT/requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy
scikit-image
scipy
matplotlib
Pillow
opencv-python
pyyaml
tensorboardX
tqdm
pandas